fix: move exception located in wrong package
옳지 않은 패키지에 위치해 있는 예외 이동
