+++ src/main/java/gift/domain/exception/GlobalExceptionHandler.java
@@ -15,8 +15,8 @@ import gift.domain.exception.notFound.MemberNotFoundException;
 import gift.domain.exception.notFound.NotFoundException;
 import gift.domain.exception.notFound.ProductNotFoundException;
 import gift.domain.exception.notFound.ProductNotIncludedInWishlistException;
-import gift.domain.exception.notFound.TokenNotFoundException;
-import gift.domain.exception.unauthorized.OptionNotFoundException;
+import gift.domain.exception.unauthorized.TokenNotFoundException;
+import gift.domain.exception.notFound.OptionNotFoundException;
 import gift.domain.exception.unauthorized.TokenUnexpectedErrorException;
 import gift.domain.exception.unauthorized.UnauthorizedException;
 import gift.global.apiResponse.ErrorApiResponse;
