+++ src/main/java/gift/domain/exception/notFound/OptionNotFoundException.java
@@ -1,0 +1,8 @@
+package gift.domain.exception.notFound;
+
+public class OptionNotFoundException extends NotFoundException {
+
+    public OptionNotFoundException() {
+        super("The option was not found.");
+    }
+}
