--- src/main/java/gift/domain/exception/unauthorized/OptionNotFoundException.java
@@ -1,8 +1,0 @@
-package gift.domain.exception.unauthorized;
-
-public class OptionNotFoundException extends UnauthorizedException {
-
-    public OptionNotFoundException() {
-        super("The option was not found.");
-    }
-}
