+++ src/main/java/gift/domain/exception/unauthorized/TokenNotFoundException.java
@@ -1,6 +1,6 @@
-package gift.domain.exception.notFound;
+package gift.domain.exception.unauthorized;
 
-public class TokenNotFoundException extends NotFoundException {
+public class TokenNotFoundException extends UnauthorizedException {
 
     public TokenNotFoundException() {
         super("토큰이 존재하지 않습니다. 발급이 필요합니다.");
