+++ src/main/java/gift/global/util/JwtUtil.java
@@ -2,7 +2,7 @@ package gift.global.util;
 
 import gift.domain.entity.Member;
 import gift.domain.exception.forbidden.TokenExpiredException;
-import gift.domain.exception.notFound.TokenNotFoundException;
+import gift.domain.exception.unauthorized.TokenNotFoundException;
 import gift.domain.exception.forbidden.TokenStringInvalidException;
 import gift.domain.exception.unauthorized.TokenUnexpectedErrorException;
 import io.jsonwebtoken.Claims;
